requirements_cloud.txt
playwright==1.42.0
rembg==2.0.53
onnxruntime==1.17.1
Pillow==10.2.0
python-dotenv==1.0.1
google-genai==0.3.0
groq==0.4.2
python-telegram-bot>=21.0
